Added Clear2 project
Created a Clear2 project that will be intended to create an executable that is similar to my current Clear2_Dev\ClearEngine. It will be able to take command line parameters and configuration files in order to determine how to run. I should be able to use this EXE just like I use my current one with my scripts.

For now, just made it able to run both RegressionTest1 and RegressionTest3. It helped me discover the common parts to these two projects.