Incorporated Clear2's Integration Strategy for FastAlign
Basically made my Clear2 code for integration of Machine models (e.g. FastAlign) compatible with how Clear3 code is written, and so I can have BuildMachineTransModel.cs and BuildTransModel.cs almost the same (modified to not write temporary sorted files).

Changed CmdLine.cs to ActionsClear3.cs to distinguish it from Actions.cs of Clear2.

Modified Program.cs to have ActionsClear3.cs output which engine is running so we can keep Program.cs the same between Clear2 and Clear3.

Updated to use RunSpe2 from the processing settings.

Updated some variables to match Clear2 in that it is not Lower but Lemma that is important for target files.