Fixed "Index out of Bound" error
Some words in the OT (Hebrew) trees will have words/lemmas that are made up of two separate words and so they are in the trees with a space. This will cause problems with segmentation of the source, and so you need to replace the space with a tilda.

I did this in my program that extracts the source text out of the trees. However, in Clear3, the source text is created on the fly from the trees, and it did not replace a space with a tilda. I changed the code to do this and then Index out of Bounds error no longer happens.