Test OT in RegressionTest1, minor changes
Added the verse file to test RegressionTest1 on OT verses to be able to see the "Index Out of Bound" error in the original Clear3 code to show to Min.  Also changed "option" to 1 in order to process from the "InputLong" subfolder.

I fixed the bug again in TreeServices.cs before making this commit.

Also just made the FastAlign enum name to mixed case to be consistent with what I have in Clear2.