Bug fix for compound words in the trees
Some words in the trees, mostly in the OT, are compound words, that is, they have a space in the middle of the word. This will cause a problem at two places.

One is during building the alignment model (alignModel) in BuildModels(). If a tree word is allowed to have a space, then the number of words and the number of word IDs will not be the same for the source since the file with IDs is created directly from the tree (not from the source text file). This will cause an "Index out of bounds" error.

Even if the word ID was created from the source text file, then there would be no "Index out of bounds" error in BuildModels(), but then there would be a bug in the tree-based auto alingment code because the number of words would not match with what is in the tree.

Replacing the space with a tilda is a hack (temporary fix). There should be a better way to represent compound words or any group of words that should be treated as a single unit (e.g. names of people, special phrases, etc.) if desired. Using a space to define the word unit, while simple, is limiting.