Bug fix for a target verse to be in a simple zone more than once.
The original code allowed a target verse to be in a SimpleZoneSpec. This does not make sense and may lead to the target side of parallel text to be extremely long.

For example, if the versifciation in a versification file to specify the versification [1-3] - [1-3] is (source - target):

1 - 1
1 - 2
1 - 3
2 - 1
2 - 2
2 - 3
3 - 1
3 - 2
3 - 3

the original code would generate [1-3] - [1-3, 2-3, 2-3]. Adding a check to see if the target verse is not already in a the list before adding it when the source verse is already in the source verse list fixes this bug.

The reason this situation was not considered (resulting in the bug) is that the original concept is that the versficiation specification for [1-3] - [1-3] would be written as a "zig-zag" mapping sequence like:

1 - 1
1 - 2
2 - 2
2 - 3
3 - 3

However, these specifications do not explicitly define what is meant by a [1-3] - [1-3] versification, and instead requires you to assume a transitive relationship in the meaning of the mappings.

The first way of expressing the versification [13] - [1-3] is created using two for loops. While this way of expressing this versification, even though explicit, is not ideal since the current algorithm for reading in a versification file, even with the bug fix, still requires that the sequence be in a particular order. For example, if the order were changed to:

1 - 1
2 - 2
3 - 3
1 - 2
1 - 3
2 - 1
2 - 3
3 - 1
3 - 2

it would result in the versification [1] - [1], [2] - [2], [3] - [3], [1] - [2, 3], [2-3] - [1-3], which is clearly not the same as [1-3] - [1-3]. So the whole topic of how to specify the versification in a file needs to be made more robust so it is not so dependent on the order of the lines in the file.