Changed Machine TokenProcessor and Alignment Heuristic
There was a redundant use of a TokenProcessor that would make the source and target lowercase. This isn't a problem with the target, but might have caused problems with the Greek since Damien said he noticed some uppercase Greek lemmas/words.

It is best to have Clear handle changing the source and target text and to not have the SMT trainer do it. Also, the GetBestAlignment will not make the source or target lowercase so there could be missed matches.

Also updated the SIL.Machine.Translation.Thot library to version 2.3.13 from 2.3.11.

Version 2.3.11 had the bug fix to prevent crashing on KHOV NT.

Version 2.3.12 exposed an interface to allow setting different heuristics for the GetBestAlignment(). The specific one I was interested in was the Intersection method (see below).

Version 2.3.13 removed doing Unicode normalization on the source and target input to the training, which caused the translation table to also have normalized text, which is different from the text Clear uses in the tree-based auto alignment. This made the results of Clear auto alignment now very close to what we got before using FastAlign.

Also, the heuristic was set to Intersection rather than the Machine default of "Och". Intersection is the same as "Min" which is what Clear2 uses with HMM.