Updated the Clear3_Tool project
The Clear2 project was changed to Clear3_Tool and was updated to behave like the current Clear2 solution.

Separate from this update, I created an identical solution with this Clear3_Tool project, and referencing the three Clear3 libraries so the library and the client solution using the libraries are separate. So eventually, this Clear3_Tool project can be removed from the Clear3 solution.