Updated to SIL library to 2.4.3, updated CLEAR.config variable names, started creating Persistence class
Needed to change some interface names to the SIL library.

Changed source text, lemma, and ID filenames to make them clearer and so also changed them in the CLEAR.config file.

Started to create persistence class to handle remembering things from one run to the next so the ClearEngine3 can be used with my scripts that run series of tests (e.g., they don't recreate parallel corpora each time.

The ClearEngine3 project is in the Clear3 solution for now so it is easier to develop. Whatever is developed will become be transferred to its own separate solution that are in ClearEngine_Run repo.