Many updates to make ClearEngine2 match with ClearEngine3 and Clear3 library
Added persistence to the source and target files of the parallel corpus by writing it to file. I have not yet written function to read it back in.

Also made getting file names for the source and target consistent with what ClearEngine2 uses in the CLEAR.config file.

Made Clear3 find trees in the same place as ClearEngine2 so I don't have to have two copies of the trees. Also made it not copy trees again if they already exist (check actual folder as opposed to only the information in the "index" file.

Updated BuildModels.cs and BuildMachineModels.cs to be the same as in ClearEngine2. They were changed to sort before writing the file that will be then read in by the Clear3 library and alignModel and transModel structures built from the files. They also include the updates to where source and target files no longer use lemma_ID within them, but have separate lemma and ID files. This was to make it consistent with what Clear3 does.

Updated ModelBuilder.cs to be the same as in ClearEngine2 where creating the transModel returns a Dictionary<string, Dictionary<string, double>> data structure instead of a Hashtable.

Updated trees in the TestSandbox1/SyntaxTrees folder to unicode normalized trees.