Modified to allow non-contiguous source verses
The previous code assumed source verses in the parallel corpus would be contiguous, that is between a first and last verse. However, some translations would map versification to non-contiguous source verses.

In reading legacy parallel corpus files, currently both Text and Lemma are instantiated to whatever is the word in the word_id file. We really don't care for how we use the parallel corpus later, but it may matter for doing the training of the SMT. So this should probably be examined more carefully.