Brought ClearEngine3 up to date with ClearEngine2
There were many changes I made to how Clear2 worked with ClearEngine2 that required me to modify ClearEngine2.

The major modification was allowing to create different SMT models from different corpora and then to use them independently in getting terminal candidates and aligning the rest (improve alignments).

This required that I modify the assumptions data structure, as well as the code that used the translation model and alignment model in auto alignment.

I also made some changes to the Source and Target.

I also added Category to SourcePoint so we are able to create lemma_cat versions of parallel corpora.

The TranslationModel now explicitly uses both SourceLemma and TargetLemma.

I also changed it so I can use C#'s CultureInfo to make more language and region lowercasing of the target in order to create lemmas.

I also changed the naming so that the target will also have lemmas rather than focus on lowercasing. This will allow us in the future to add a way to actually get lemmas from surface forms for the target.

I also changed the filtering of words from parallel corpora to be generic rather than it only being used to filter out function words.

Also made importing assumptions not require that the files exist. If they don't, then just create an empty data structure (i.e. read in an empty file). This is to match Clear2 in not requiring there to be these files.

Updated the build model files to match what we now do with the scripts.

Updated the regression and unit tests to work with the modified interfaces.

The last big change was to the Actions.cs and Program.cs files in the Clear3 project.