Added reusing token, corpora, and model files
Made Clear3 consistent with Clear2 in allowing reuse of files based upon certain flags. These flags can be set from the command line so scripts can create new files upon the first run, but then reuse them later. It is probably best to have script erase all the files as well since a file needed later might not have been always created by a previous alignment, in which case, you can have it set to reuse files from the very beginning.

In order to reuse files, added importing token, translation model, and alignment model.

There are other changes in this commit that are not significant since they are code added to help me debug an index out of bound error. I isolated to the fact that creating terminal candidates sometimes has a candidate with a "Not a Number" value for its LogScore, which later causes a candidate to be eliminated, which causes the index out of bound error. I have not yet figured out where the problem exactly occurs and so I have not fixed this bug yet in this commit.