Commit to GitHub so I can move over to my Mac and continue the work to sync with Clear2/ClearEngine2
This is a inconsistent version where there are still errors and missing code. I'm only committing it so I can move over to my Mac, which has two screens to sync up the Clear3/ClearEngine3 code with the Clear2/ClearEngine2 code.

It does have the new implementation of BuildGizaTransModels.cs and BuildTransModel.cs that I just modified for C/CE2 so that a string for the path to the Python executable file can be passed in rather than being hardcoded.

Eventually, it may make more sense that since this is only specific to this implementation of IBM4 using Giza++ that this should be handled internally (i.e., it can have it's own configuration file in the giza-py folder).