Synced Clear/ClearEngine3 with Clear/ClearEngine2 and successfully compiled but not yet tested
In an earlier incomplete commit (in order to migrate the repo to ClearBible), it had the update to allow for a lemmatization stage that allows the use of morphological analysis to break up words into morphemes. This is important for not only to do better alignment with agglutinative languages but it also helps when there is a mismatch between whether a language uses case or whether it uses prepositions.

Refactored the building of the parallel corpora and the models.

Updated the command line processing.

Updated the actions to use delegate to allow lists of methods that can be run sequentially.

Changed puncs, stopWords, sourceFunctionWords, and targetFunctionWords from List<string> to HashSet<string> since HashSet is a better structure fo these since all we are doing is seeing if something is in the set (order is not important).

Added some methods to get LpaLemmaLine from alignments since we want to save an alignment file that is based upon the lemma/morphemes and then create a legacy alignment (LpaLine) file that is based upon words.

Also partially added the code to allow for specifying the path to the Python executable in the Clear.config file so the Giza++ implementation can be used without having to recompile since Mac and Windows location of the file are different. Because ClearEngine3 is much better in terms of the interface design, I think passing in the location for just the Giza++ SMT model building doesn't make sense. This should probably be taken care of internal to that part, and so I am thinking of changing this, so that is why I only partially implemented it.